# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GeoPagos Reto Api. Edwin Jose Mogollon Escalona eJme1992",
        "description": "Documentación de la API para la gestión de torneos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tournaments/register": {
            "post": {
                "tags": [
                    "Tournament"
                ],
                "summary": "Crear un nuevo torneo",
                "description": "Registers a new player with the specified details. Requires JWT token for authentication.",
                "operationId": "fee228ea0319b6d99ffc736c8f433d64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "gender",
                                    "name",
                                    "number_players"
                                ],
                                "properties": {
                                    "gender": {
                                        "type": "string",
                                        "example": "male"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Torneo de Verano"
                                    },
                                    "number_players": {
                                        "type": "integer",
                                        "example": 16
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Torneo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los parámetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El número de participantes debe ser mayor a cero y múltiplo de dos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tournaments/register-player": {
            "post": {
                "tags": [
                    "Tournament"
                ],
                "summary": "Registrar un jugador en un torneo",
                "description": "Registers a new player with the specified details. Requires JWT token for authentication.",
                "operationId": "7b877d3ae80a62eed4b31c27f001972f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tournament_id",
                                    "player_id"
                                ],
                                "properties": {
                                    "tournament_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "player_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jugador registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los parámetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El torneo ya tiene el número de jugadores requeridos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tournaments/is-complete/{tournamentId}": {
            "get": {
                "tags": [
                    "Tournament"
                ],
                "summary": "Verificar si el torneo está completo",
                "description": "Registers a new player with the specified details. Requires JWT token for authentication.",
                "operationId": "52f8e1617881d805446686398e365eef",
                "parameters": [
                    {
                        "name": "tournamentId",
                        "in": "path",
                        "description": "ID del torneo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado del torneo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isComplete": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tournaments/gender/{gender}": {
            "get": {
                "tags": [
                    "Tournament"
                ],
                "summary": "Obtener torneos por género",
                "description": "Registers a new player with the specified details. Requires JWT token for authentication.",
                "operationId": "ab83993755b2b8c1bad6b4c718ef0e2e",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "path",
                        "description": "Género de los torneos a filtrar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "male"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de torneos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Torneo de Verano"
                                                    },
                                                    "gender": {
                                                        "type": "string",
                                                        "example": "male"
                                                    },
                                                    "number_players": {
                                                        "type": "integer",
                                                        "example": 16
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}